minor fixes to how functions are called
Unfortunately, due to JS weirdness, wrapping the navigateTo() function in a method in the abstract class does not work as straighforwardly for eventlisteners.
In an attempt to keep using the functions still in a relative consitent way (using this.) the eventhadnlers HAVE TO be declared as arrow functions.
Look into the scope of this and arrow functions, if you're interested